refactored out preferences to reuse for pageloadsetting
