+++ views/popup.html
@@ -59,6 +59,10 @@
     </div>
   </div>
 
+  <button id="onPageLoadBtn" class="w-100 flex flex-column align-items-center">
+    On PageLoad
+  </button>
+
   <button id="resetDefaultsBtn" class="w-100 flex flex-column align-items-center">
     Reset Defaults
   </button>
